+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Added support for models from package *lcmm* (`lcmm`, `externVar`).
 
+* Support for objects from the *modelbased* package. This includes methods for
+  `get_parameters()`, `get_statistic()`, `get_df()`, `get_varcov()`,
+  `find_formula()`, `find_parameters()`, `find_statistic()`.
+
 * Better formatting for p-values-columns from equivalence tests from other
   packages.